Release 1.0.0  (#49)
* fix(): Member, Reservation 연관관계 ManyToOne, Seat Reservation연관관계 OneToOne으로 변경

* feat(): 1분(추후에 15분으로 변경)안에 결제가 되지 않을 시 예약좌석 lock 풀기 및 예약 삭제

* refactor(): custoemException 적용

* Update ErrorCode.java

* fix(): 슬래시 삭제

* fix: logout deprecated 삭제

* refactor: 결제 완료시 티켓에 결제 버튼 숨김

* refactor: 티켓 리스트 형식으로 출력, 결제 실패시 /ticket으로 이동

* test: 테스트 수정

* refactor(): 예약이 정상적으로 진행되지 않은 경우 현재 페이지 새로고침으로 변경

* feat(): 예약 취소 구현

* chore: pinpoint 환경 구성을 위해 endpoint 수정

* chore: pinpoint 환경 구성을 위해 endpoint 수정

* chore: pinpoint 사용하지 않으므로 빌드 파일 수정

* chore: elastic apm을 위한 보안 규칙 수정

* chore: elastic apm을 위해 docker file 수정 및 agent 파일 추가

* fix: apm server 로그인 처리를 위해 로그인 필터 uri 변경

* fix: /checklogin uri 추가

* fix: /checklogin uri 다시 제거

* fix: cors 설정 추가

* fix: http host 사용하지 않도록 수정

* chore: logging 추가

* fix: cors 문제 해결하도록 수정

* fix: logging 삭제

* feat: release, dev 브랜치에 따라 다르게 deploy 되도록 작성

* fix: gradle 조건문 수정

* fix: eip로 ip 주소 수정

* fix(): api 엔드 포인트 수정 (api->app-api)

* refactor(): 예약 취소 - 예약자 확인 로직 추가

* feat(): 테스트를 위한 MemberGenerator 생성

* fix(): app-api 경로 이름 삭제 (#52)

* fix: dev, release 별로 application.yml 다르게 적용되도록 수정 (#51)

---------

+++ src/main/java/com/example/pigonair/domain/reservation/service/ReservationServiceImpl.java
@@ -71,8 +71,13 @@ public class ReservationServiceImpl implements ReservationService {
     @Transactional
     public void cancelReservation(Long reservation_id, UserDetailsImpl userDetails) {
         Reservation reservation = getReservation(reservation_id);
-        reservation.getSeat().setIsAvailable();
-        reservationRepository.delete(reservation);
+        Member member = getMember(userDetails);
+        if(member.equals(reservation.getMember())){
+            reservation.getSeat().setIsAvailable();
+            reservationRepository.delete(reservation);
+        }
+        else
+            throw new CustomException(ErrorCode.FORBIDDEN);
     }
